requirements.txt
python-telegram-bot==20.7
cohere==4.37
python-dotenv==1.0.0
flask==2.3.3 